rag_app/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
fastapi-cors==0.0.6

# PDF processing
pymupdf==1.23.8

# Text processing and chunking
tiktoken==0.5.2

# Vector store
faiss-cpu==1.7.4

# Reranking
sentence-transformers==2.7.0
torch>=2.0.0; platform_system!="Windows"
huggingface-hub>=0.20.0

# Data processing
pandas==2.1.4
pyarrow==14.0.1

# Database
sqlite-utils==3.35.2

# OpenAI
openai>=1.40.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and utilities
python-json-logger==2.0.7
